Moved more utility functions into Context implementation
